chore: upgrade to .NET 8 and update NuGet packages
- Upgrade TargetFramework from net6.0-windows to net8.0-windows10.0.17763.0
- Upgrade Tests from net5.0 to net8.0
- Update Serilog packages (2.12.0→4.3.0, Extensions.Hosting 5.0.1→8.0.0, etc.)
- Update Microsoft.* packages for .NET 8 compatibility
  - CsWin32 0.1.588-beta→0.3.264
  - VisualStudio.Threading 17.5.22→17.14.15
  - System.Management/ServiceController/CodePages 7.0.0→8.0.x
- Update Fody 6.6.4→6.9.1, MaxMind.GeoIP2 5.1.0→5.2.0
- Update test packages (MSTest 3.0.2→3.7.0, coverlet 3.2.0→6.0.2)
- Fix CsWin32 0.3.x breaking changes:
  - RouteHelper: ADDRESS_FAMILY type changes, NTSTATUS API updates
  - PortHelper: GetExtendedTcpTable signature change
  - MainForm: HWND constructor disambiguation

+++ Netch/Forms/MainForm.cs
@@ -397,7 +397,7 @@ public partial class MainForm : Form
 
     private void ShowHideConsoleToolStripMenuItem_Click(object sender, EventArgs e)
     {
-        var windowStyles = (WINDOW_STYLE)PInvoke.GetWindowLong(new HWND(Program.ConsoleHwnd), WINDOW_LONG_PTR_INDEX.GWL_STYLE);
+        var windowStyles = (WINDOW_STYLE)PInvoke.GetWindowLong(new HWND((nint)Program.ConsoleHwnd), WINDOW_LONG_PTR_INDEX.GWL_STYLE);
         var visible = windowStyles.HasFlag(WINDOW_STYLE.WS_VISIBLE);
         PInvoke.ShowWindow(Program.ConsoleHwnd, visible ? SHOW_WINDOW_CMD.SW_HIDE : SHOW_WINDOW_CMD.SW_SHOWNOACTIVATE);
     }
